refactor(fetcher): Move fecthPairData form Pair class to Fetcher class
